@@ -1,3 +1,3 @@
-import * as Rx from "rxjs/Observable";
+import * as Rx from "rxjs";
 
 console.log(Rx);
